requirements.txt
django
django-jinja
django-magiclink